refactor: Move home_page.dart into app dir
