+++ lib/app/app.dart
@@ -1,6 +1,6 @@
 import 'package:confwebsite2023/core/config.dart';
 import 'package:confwebsite2023/core/theme/app_theme.dart';
-import 'package:confwebsite2023/home_page.dart';
+import 'package:confwebsite2023/app/home_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
